requirements.txt
asttokens==2.4.1
attrs==23.2.0
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.1.3
cloudpickle==1.6.0
cycler==0.11.0
decorator==5.1.1
defusedxml==0.7.1
docker-pycreds==0.4.0
docopt==0.6.2
executing==2.0.1
fastjsonschema==2.20.0
fonttools==4.33.3
future==0.18.2
gitdb==4.0.9
GitPython==3.1.27
gym==0.19.0
gym_gridverse==0.0.1
gym-notices==0.0.8
idna==3.3
imageio==2.9.0
imageio-ffmpeg==0.5.0
importlib_metadata==7.1.0
importlib_resources==6.4.0
ipython==8.12.3
jedi==0.19.1
Jinja2==3.1.4
joblib==1.1.0
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
kiwisolver==1.4.2
MarkupSafe==2.1.5
matplotlib==3.5.2
matplotlib-inline==0.1.7
mistune==3.0.2
more-itertools==10.2.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
numpy==1.24.4
packaging==21.3
pandas==1.4.2
pandocfilters==1.5.1
parso==0.8.4
pathtools==0.1.2
pexpect==4.9.0
pickleshare==0.7.5
Pillow==7.2.0
pip==24.0
pipreqs==0.5.0
pkgutil_resolve_name==1.3.10
platformdirs==4.2.2
ply==3.11
promise==2.3
prompt_toolkit==3.0.47
protobuf==3.20.1
psutil==5.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
pyglet==1.5.0
Pygments==2.18.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
pyzmq==26.0.3
referencing==0.35.1
requests==2.27.1
rpds-py==0.18.1
schema==0.7.7
scipy==1.8.1
seaborn==0.11.2
sentry-sdk==1.5.12
setproctitle==1.2.3
setuptools==69.5.1
shortuuid==1.0.9
six==1.16.0
smmap==5.0.0
soupsieve==2.5
stack-data==0.6.3
tinycss2==1.3.0
torch==1.13.1+cpu
torchaudio==0.13.1+cpu
torchvision==0.14.1+cpu
tornado==6.4.1
traitlets==5.14.3
typing_extensions==4.2.0
urllib3==1.26.9
wandb==0.12.16
wcwidth==0.2.13
webencodings==0.5.1
wheel==0.43.0
yarg==0.1.9
zipp==3.19.1
